+++ petl/io/codec/__init__.py
@@ -1,7 +1,0 @@
-from __future__ import absolute_import, print_function, division
-
-from petl.io.codec.zstd import ZstandardCodec
-
-from petl.io.codec.lz4 import LZ4Codec
-
-from petl.io.codec.xz import XZCodec