+++ petl/io/source/__init__.py
@@ -1,5 +1,0 @@
-from __future__ import absolute_import, print_function, division
-
-from petl.io.source.s3 import S3Source
-
-from petl.io.source.smb import SMBSource